summarization/youtube-summarizer/requirements.txt
google-generativeai
youtube-transcript-api
yt-dlp
python-dotenv
pydantic